224 deactivated users are deleted from the system (#75)
* Removed action domain parameters.

It seems that when users are inactive they do not count as belonging to any particular category, as the extra domain parameter for the categories filters them out only when they are inactive. I've removed the category domain because it looks for all categories anyway, so should not have any affect on the results (except inactive users of course).

* - Refactoring user management and associated tests
- Adding test for EOBS-224

* - Adding Doc Strings as per @Nayira 's request
