+++ nh_clinical/data/change_ward_manager_to_shift_coordinator.xml
@@ -57,9 +57,7 @@
             <field name="name">Account Administration</field>
             <field name="res_model">nh.clinical.user.management</field>
             <field name="view_mode">tree,form</field>
-            <field name="domain">[['category_ids.name','in',
-                ['HCA','Nurse','Shift Coordinator','Doctor','Senior Manager','Kiosk','System Administrator','Senior Doctor',
-                'Junior Doctor','Registrar','Consultant','Receptionist']],['login','not in',['adt','admin']]]</field>
+            <field name="domain">[['login','not in',['adt','admin']]]</field>
             <field name="search_view_id" ref="view_nhc_user_management_search"/>
             <field name="context">{'partner_category_display': 'short'}</field>
         </record>
