+++ nh_clinical/models/__init__.py
@@ -1,0 +1,1 @@
+from . import user_management
