+++ nh_clinical/tests/__init__.py
@@ -9,7 +9,6 @@ from . import test_location
 from . import test_operations
 from . import test_patient_placement_wizard
 from . import test_responsibility_allocation_wizard
-from . import test_user_management
 from . import test_users
 
 from .nh_clinical_doctor_allocation import *
@@ -17,6 +16,7 @@ from .nh_clinical_patient import *
 from .nh_clinical_staff_allocation import *
 from .nh_clinical_staff_reallocation import *
 from .res_user import *
+from .nh_clinical_user_management import *
 
 # Disabled Tests
 # from . import test_doctor
