+++ nh_clinical/tests/nh_clinical_user_management/__init__.py
@@ -1,0 +1,3 @@
+from . import test_user_crud
+from . import test_views_actions
+from . import test_activate_deactivate
